# src/views/pages/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": ".M.U.F.A.R. APIs",
    "version": "1.3.8",
    "x-logo": {
      "url": "/assets/img/logo.png",
      "backgroundColor": "#000",
      "altText": "Example logo"
    }
  },
  "security": [
    {
      "apiKey": []
    }
  ],
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "query",
        "name": "apikey"
      }
    },
    "schemas": {
      "SuccessResponse": {
        "description": "Success"
      }
    }
  },
  "paths": {
...
    "/pinterest/getBoard": {
      "get": {
        "tags": [
          "Pinterest"
        ],
        "description": "Retrieve data for a specific board",
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the board"
          },
          {
            "in": "query",
            "name": "slashurl",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The slash URL of the board"
          },
          {
            "in": "query",
            "name": "bookmark",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Bookmark parameter for pagination"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
    "/pinterest/searchBoards": {
      "get": {
        "tags": [
          "Pinterest"
        ],
        "description": "Search for boards based on query",
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The query string for searching boards"
          },
          {
            "in": "query",
            "name": "bookmark",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Bookmark parameter for pagination"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
    "/pinterest/searchPins": {
      "get": {
        "tags": [
          "Pinterest"
        ],
        "description": "Search for pins based on query",
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The query string for searching pins"
          },
          {
            "in": "query",
            "name": "bookmark",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Bookmark parameter for pagination"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
    "/pinterest/suggestions": {
      "get": {
        "tags": [
          "Pinterest"
        ],
        "description": "Get suggestions for related pins based on pin ID",
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the pin"
          },
          {
            "in": "query",
            "name": "bookmark",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Bookmark parameter for pagination"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}